project/requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
pdfplumber==0.5.28
python-docx==0.8.11
easyocr==1.3.2
SQLAlchemy==1.4.22
scikit-learn==0.24.2
python-dateutil==2.8.1
Werkzeug==2.0.2
pyodbc==4.0.32
spacy==3.1.3
flask-swagger-ui==3.36.0
Cython==0.29.21
